+++ src/test/java/com/mysite/sbb/legaldong/LegalDongMigrationServiceTests.java
@@ -24,7 +24,8 @@ public class LegalDongMigrationServiceTests {
 		LegalDongExcelParser parser = new LegalDongExcelParser();
 		LegalDongMigrationService service = new LegalDongMigrationService(
 				parser,
-				new LegalDongJdbcUpsertRepository(null));
+				new LegalDongJdbcUpsertRepository(null),
+				null);
 
 		LegalDongMigrationPreviewResult preview = service.preview(xlsx, 0, 20);